adjusted requested_usd model field type

+++ game/models.py
@@ -53,7 +53,7 @@ class TopUps(models.Model):
 
 class Payouts(models.Model):
     player = models.ForeignKey(Players, on_delete=models.CASCADE)
-    requested_usd = models.IntegerField()
+    requested_usd = models.FloatField()
     calculated_eth = models.FloatField(default=0)
     paid = models.BooleanField(default=False)
     created = models.DateTimeField(auto_now=True)
